+++ test/services/population_service_test.dart
@@ -75,6 +75,7 @@ void main() {
       ]);
 
       final actual = await testObject.reproduce(
+        populationSize: populationSize,
         population: population,
         wave: wave,
       );